requirements.txt
mysql-connector-python==8.0.33
configparser==6.0.0
